restored original threaded example
and made a new one to show how to change pygment styles
